+++ src/storage/kafka.go
@@ -4,9 +4,12 @@ import (
 	"context"
 	"encoding/json"
 	"log/slog"
+	"os"
 
 	"github.com/segmentio/kafka-go"
-	"github.com/shyam0507/pd-order/types"
+	"github.com/segmentio/kafka-go/sasl/plain"
+	"github.com/shyam0507/pd-order/src/src/types"
+
 	"go.mongodb.org/mongo-driver/bson/primitive"
 )
 
@@ -62,22 +65,45 @@ func (k KafkaProducer) ProduceOrderConfirmed(key string, value types.OrderConfir
 }
 
 func NewKafkaProducer(topic string, brokers []string) Producer {
+	username := os.Getenv("KAFKA_USERNAME")
+	password := os.Getenv("KAFKA_PASSWORD")
+	mechanism := plain.Mechanism{
+		Username: username,
+		Password: password,
+	}
+
 	w := &kafka.Writer{
 		Addr:     kafka.TCP(brokers...),
 		Topic:    topic,
 		Balancer: &kafka.LeastBytes{},
+		Transport: &kafka.Transport{
+			SASL: mechanism,
+		},
+		AllowAutoTopicCreation: true,
 	}
 
 	return &KafkaProducer{writer: w}
 }
 
 func NewKafkaConsumer(topic string, brokers []string, storage Storage, producer Producer) Consumer {
+	username := os.Getenv("KAFKA_USERNAME")
+	password := os.Getenv("KAFKA_PASSWORD")
+	mechanism := plain.Mechanism{
+		Username: username,
+		Password: password,
+	}
+
+	dialer := &kafka.Dialer{
+		SASLMechanism: mechanism,
+	}
+
 	r := kafka.NewReader(kafka.ReaderConfig{
 		Brokers:     brokers,
 		Topic:       topic,
 		GroupID:     "pd-order-group",
 		StartOffset: kafka.LastOffset,
 		MaxBytes:    10e6, // 10MB
+		Dialer:      dialer,
 	})
 
 	return KafkaConsumer{reader: r, storage: storage, producer: producer}
